+++ WestWind.Globalization.Sample.AspNetCore/Properties/Resources.Designer.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 
-namespace WestWind.Globalization.Sample.AspNetCore.Resources {
+namespace WestWind.Globalization.Sample.AspNetCore.Properties {
     using System;
     
     
@@ -39,7 +39,7 @@ namespace WestWind.Globalization.Sample.AspNetCore.Resources {
         internal static global::System.Resources.ResourceManager ResourceManager {
             get {
                 if (object.ReferenceEquals(resourceMan, null)) {
-                    global::System.Resources.ResourceManager temp = new global::System.Resources.ResourceManager("WestWind.Globalization.Sample.AspNetCore.Resources.Resources", typeof(Resources).Assembly);
+                    global::System.Resources.ResourceManager temp = new global::System.Resources.ResourceManager("WestWind.Globalization.Sample.AspNetCore.Properties.Resources", typeof(Resources).Assembly);
                     resourceMan = temp;
                 }
                 return resourceMan;