+++ Westwind.Globalization.AspnetCore/WestwindGlobalizationConfiguration.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
+using Westwind.Utilities;
 
 namespace Westwind.Globalization
 {
@@ -15,9 +16,10 @@ namespace Westwind.Globalization
     public static class ServiceCollectionExtensions
     {
         public static IServiceCollection  AddWestwindGlobalization(this IServiceCollection services, Action<DbResourceConfiguration> setOptionsAction)
-        {           
+        {
             DbResourceConfiguration.Current.Initialize();
             setOptionsAction.Invoke(DbResourceConfiguration.Current);            
+
             return services;
         }
     }